src/requirements.txt
Flask==2.0.1
flask-restx==0.5.1
flask-cors==3.0.10
google-auth==2.3.3
google-cloud-sql==0.4.0
mysql-connector-python==8.0.26
gunicorn==20.1.0